requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
pytest-playwright>=0.4.0
apscheduler>=3.10.0
pydantic>=2.0.0
httpx>=0.25.0
